Create subspecs 🍕
At long last, here are the first subspecs of Alicerce!
Sorry for the delay 😇

Essentially, the subspecs were a mapping of our already existing
internal folder structure, with some exceptional groupings.

Currently defined subspecs are:

- Core (Extensions, Utils and Shared)

- Analytics

- Deeplinking

- Logging

- Network

- UI (Observers, QuartzCore, UIKit)

- PerformanceMetrics

- Persistence

- Resource

- Stores

- View

Closes #122.
