+++ Sources/Network/Resource/AuthenticatedRequestResource.swift
