+++ Sources/Network/Resource/HTTPNetworkResource.swift
