+++ Sources/Network/Resource/HTTPResourceEndpoint.swift
