+++ Sources/Network/Resource/NetworkResource.swift
