+++ Tests/AlicerceTests/View/CollectionViewCellTestCase.swift
