+++ Tests/AlicerceTests/View/TestNIBCollectionReusableView.swift
@@ -1,4 +1,5 @@
 import UIKit
+import Alicerce
 
 class TestNIBCollectionReusableView: CollectionReusableView, NibView {
     @IBOutlet weak var testLabel: UILabel!
