+++ Tests/AlicerceTests/View/TestNIBCollectionViewCell.swift
@@ -1,4 +1,5 @@
 import UIKit
+import Alicerce
 
 class TestNIBCollectionViewCell: UICollectionViewCell, NibView {
     @IBOutlet weak var testLabel: UILabel!
