+++ Tests/AlicerceTests/View/TestNIBTableHeaderFooterView.swift
@@ -1,4 +1,5 @@
 import Foundation
+import Alicerce
 
 class TestNIBTableHeaderFooterView: UITableViewHeaderFooterView, NibView {
     @IBOutlet weak var testLabel: UILabel!
