+++ Tests/AlicerceTests/View/TestNIBTableViewCell.swift
@@ -1,4 +1,5 @@
 import UIKit
+import Alicerce
 
 class TestNIBTableViewCell: UITableViewCell, NibView {
     @IBOutlet weak var testLabel: UILabel!
